+++ src/main/java/com/telepathicgrunt/blame/mixin/BootstrapMixin.java
@@ -3,7 +3,9 @@ package com.telepathicgrunt.blame.mixin;
 import com.telepathicgrunt.blame.main.DispenserBlockRegistry;
 import net.minecraft.Bootstrap;
 import net.minecraft.block.dispenser.DispenserBehavior;
+import net.minecraft.block.dispenser.ItemDispenserBehavior;
 import net.minecraft.item.Item;
+import net.minecraft.util.Util;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -22,8 +24,9 @@ public class BootstrapMixin {
 			at = @At(value = "TAIL"))
 	private static void onInit(CallbackInfo ci)
 	{
-		DispenserBlockRegistry<Item, DispenserBehavior>  map = new DispenserBlockRegistry<>();
+		DispenserBlockRegistry<Item, DispenserBehavior>  map = Util.make(new DispenserBlockRegistry<>(), (behaviour) -> behaviour.defaultReturnValue(new ItemDispenserBehavior()));
 		map.putAll(DispenserBlockAccessor.getBEHAVIORS());
+		map.startupIgnore = false; // Finished copying. Now turn on registry replacement detection.
 		DispenserBlockAccessor.setBEHAVIORS(map);
 	}
 }